pyproject.toml
[build-system]
requires = ["setuptools>=61.0","wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "socmed-poster"
version = "0.1.0"
description = "Unified poster for Facebook, Twitter, Instagram, and LinkedIn with a Flask UI and CLI clients."
readme = "README.md"
authors = [ { name = "SeanSo-hub" } ]
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "flask>=2.3.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "tweepy>=4.14.0",
    "Pillow>=10.0.0"
]

[project.urls]
"Homepage" = "https://example.invalid/"

[project.scripts]
socmed-poster = "socmed_poster.web:create_app"
